+++ changelog.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+- Fix panic when no docker daemon is running (#457)
 - Remove unused journal logging (#431)
 
 ## v1.0.1302 (2018-07-26)