+++ changelog.md
@@ -1,6 +1,7 @@
 ## unreleased
 
 - Clean the docker box option when cleaning the network (#462)
+- RDD Cleanup for errors/aborts (#461)
 
 ## v1.0.1302 (2018-07-26)